requirements.txt
customtkinter
yt-dlp
pyglet
Pillow
mutagen